+++ SuperMarioBros/SuperMarioBros/PlayerCharacter/PlayerStates/LeftJumpingPlayerState.cs
@@ -11,7 +11,7 @@ namespace SuperMarioBros.PlayerCharacter.PlayerStates
     public class LeftJumpingPlayerState : AbstractPlayerState, IJumpingPlayerState, ILeftFacing
     {
         private int fallingSpeed;
-        public LeftJumpingPlayerState(Player player, int jumpingSpeed = 140) : base(player)
+        public LeftJumpingPlayerState(Player player, double jumpingSpeed = 140) : base(player)
         {
             player.Sprite = PlayerSpriteFactory.Instance.CreateLeftJumpingPlayerSprite();
             JumpingSpeed = jumpingSpeed;
