+++ SuperMarioBros/SuperMarioBros/PlayerCharacter/PlayerStates/LeftMovingPlayerState.cs
@@ -12,7 +12,7 @@ namespace SuperMarioBros.PlayerCharacter.PlayerStates
     public class LeftMovingPlayerState : AbstractPlayerState, ILeftFacing
     {
         private bool stop;
-        public LeftMovingPlayerState(Player player, int speed = -20) : base(player)
+        public LeftMovingPlayerState(Player player, double speed = -20) : base(player)
         {
             player.Sprite = PlayerSpriteFactory.Instance.CreateLeftMovingPlayerSprite();
             if (speed > 0)
